+++ src/common/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './encryptionData';
+export * from './isEnc';
 export * from './parseKeyBitLength';